Add default HttpClient to app/build.gradle (#97)
A default HttpClient implementation should be added to build.gradle, or there will be an issue "A request was made to load the default HttpClient provider but one could not be found on the classpath. If you are using a dependency manager, consider including a dependency on azure-core-http-okhttp or azure-core-http-httpurlconnection. Depending on your existing dependencies, you have the choice of OkHttp or HttpUrlConnection implementations. Additionally, refer to https://aka.ms/azsdk/java/docs/custom-httpclient to learn about writing your own implementation."